\name{forestSAS_simapp}
\alias{forestSAS_simapp}
\title{
  App for forestSAS package is used for simulated data
}
\details{
Forest structure commonly refers to a distribution pattern of tree attributes within a forest ecosystem. Similarly, tree population structure describes the distribution characteristics of individuals of conspecifics within a community, and the spatial structure of a tree population is largely determined by the relationships among neighboring groups of trees.
}
\author{
Zongzheng Chai, chaizz@126.com
}
\references{
None
}
\examples{
forestSAS_simapp()
}
