\name{tree.ppp}
\alias{tree.ppp}
\docType{data}
\title{
Sample data for analizing the forest spatial structure.
}
\description{
A sample data of field survey, to help us understanding the method of forest spatial structure analysis.class is "ppp".
}
\usage{data("tree.ppp")}
\details{
The function "ppp" in the package spatstat
}
\references{
None
}
\examples{
data(tree.ppp)
}

